backend/requirements.txt
Django==4.2.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.3.0
django-storages==1.14.2
psycopg2-binary==2.9.9
redis==5.0.0
celery==5.3.4
pydicom==2.4.1
Pillow==10.1.0
requests==2.31.0
boto3==1.34.2
python-decouple==3.8
gunicorn==21.2.0
django-environ==0.11.2
daphne==4.1.2
django-redis==5.4.0
